+++ components/TopToolbar.tsx
@@ -28,13 +28,9 @@ const ViewRadioButton: React.FC<RadioProps> = ({
 );
 
 const descriptions: Record<CollectionView, string> = {
-  [CollectionView.Simple]:
-    "Track all the decors which the Pikmin Bloom game tracks with the Simple view.",
-  [CollectionView.Advanced]:
-    "Track all the decors which the Pikmin Bloom game tracks, plus roadside stickers, with the Advanced view.",
-  [CollectionView.Exhaustive]:
-    "Track all the decors which the Pikmin Bloom game tracks, plus roadside stickers in individual colors, with the Exhaustive view.",
-};
+  [CollectionView.Mitchell]:
+    "Track all the decors which the Pikmin Bloom game tracks, plus roadside stickers, with the Mitchell view.",
+  };
 
 export const TopToolbar: React.FC = () => {
   const { view, setView } = useViewContext();
@@ -44,27 +40,17 @@ export const TopToolbar: React.FC = () => {
 
   return (
     <>
-      <div>
+      {/* <div>
         View:
+        
         <ViewRadioButton
-          label="Simple"
+          label="Mitchell"
           view={view}
-          value={CollectionView.Simple}
+          value={CollectionView.Mitchell}
           onChange={onChange}
         />
-        <ViewRadioButton
-          label="Advanced"
-          view={view}
-          value={CollectionView.Advanced}
-          onChange={onChange}
-        />
-        <ViewRadioButton
-          label="Exhaustive"
-          view={view}
-          value={CollectionView.Exhaustive}
-          onChange={onChange}
-        />
-      </div>
+        
+      </div> */}
       <p>{descriptions[view]}</p>
     </>
   );